\name{grasp.adm.dump}
\alias{grasp.adm.dump}
\title{ Internal GRASP function }
\description{
This function saves the current version of your GRASP functions in a dump file that can be transported to another R session or to another computer.
}
\usage{
grasp.adm.dump()
}
\author{ Anthony.Lehmann@unige.ch }
\seealso{ grasp \code{\link{grasp}}, grasp.in \code{\link{grasp.in}}}
\keyword{models}
