\name{grasp.data.read}
\alias{grasp.data.read}
\title{ Internal GRASP function }
\description{
Read and load YYY,XXX, XXXpred text tables from disk
}
\usage{
grasp.data.read(path="")
}
\arguments{
  \item{path}{The path (directory) in which the files are located}
}
\author{ Anthony.Lehmann@unige.ch }
\seealso{ grasp \code{\link{grasp}}, grasp.in \code{\link{grasp.in}}}
\keyword{models}