\name{grasp.exp.summary}
\alias{grasp.exp.summary}
\title{ Internal GRASP function }
\description{
grasp.exp.summary prints summaries of XXX and YYY in order to check for missing data (NA) and distribution of min and max values of predictors and responses.
}
\usage{
grasp.exp.summary(gr.Yi)
}
\arguments{
  \item{gr.Yi}{A vector containing the selected responses}
}

\author{ Anthony.Lehmann@unige.ch }
\seealso{ grasp \code{\link{grasp}}, grasp.in \code{\link{grasp.in}}}
\keyword{models}