\name{grasp.int.contplot}
\alias{grasp.int.contplot}
\title{ Internal GRASP function }
\description{
grasp.cont.plot() allows plotting of the drop, model and alone contributions calculated by grasp.contrib() and saved in DROP.CONTRIB, MODEL.CONTRIB and ALONE.CONTRIB dataframes.
}
\usage{
grasp.int.contplot(gr.Yi)
}
\arguments{
  \item{gr.Yi}{A vector containing the selected responses}
}
\author{ Anthony.Lehmann@unige.ch }
\seealso{ grasp \code{\link{grasp}}, grasp.in \code{\link{grasp.in}}}
\keyword{models}