\name{grasp.pred.plot}
\alias{grasp.pred.plot}
\title{ Internal GRASP function }
\description{
This function maps within R the predictions made by grasp.pred() and stored in gr.predmat.
}
\usage{
grasp.pred.plot(gr.Yi)
}
\arguments{
  \item{gr.Yi}{A vector containing the selected responses}
}
\details{
The resolution of the exported GRID can be set with the resolution option.
}
\author{ Anthony.Lehmann@unige.ch }
\seealso{ grasp \code{\link{grasp}}, grasp.in \code{\link{grasp.in}}}
\keyword{models}