\name{grasp.step.crossplot}
\alias{grasp.step.crossplot}
\title{ Internal GRASP function }
\description{None}
\usage{
grasp.step.crossplot(gr.Yi)
}
\arguments{
  \item{gr.Yi}{A vector containing the selected responses}
}
\author{ Anthony.Lehmann@unige.ch }
\seealso{ grasp \code{\link{grasp}}, grasp.in \code{\link{grasp.in}}}
\keyword{models}