\name{grasp.step.crossval}
\alias{grasp.step.crossval}
\title{ Internal GRASP function }
\description{None}
\usage{
grasp.step.crossval(tested.model)
}
\arguments{
  \item{tested.model}{...}
}
\author{ Anthony.Lehmann@unige.ch }
\seealso{ grasp \code{\link{grasp}}, grasp.in \code{\link{grasp.in}}}
\keyword{models}