\name{grasp.step.rocplot}
\alias{grasp.step.rocplot}
\title{Internal GRASP function }
\description{None}
\usage{
grasp.step.rocplot(z1, z2)
}
\arguments{
  \item{z1}{...}
  \item{z2}{...}
}
\author{ Anthony.Lehmann@unige.ch }
\seealso{ grasp \code{\link{grasp}}, grasp.in \code{\link{grasp.in}}}
\keyword{models}