\name{grasp.step.scope}
\alias{grasp.step.scope}
\title{ Internal GRASP function }
\description{
  grasp.step.scope() creates a model scope for the stepwise selection of variables with all selected Predictor Variables including factors and smoothed continuous variables, using both degree of freedom (DF1 and DF2)
}
\usage{
grasp.step.scope(sX)
}
\arguments{
  \item{sX}{...}
}
\author{ Anthony.Lehmann@unige.ch }
\seealso{ grasp \code{\link{grasp}}, grasp.in \code{\link{grasp.in}}}
\keyword{models}