\name{grasp.step.start}
\alias{grasp.step.start}
\title{Internal GRASP function }
\description{
  grasp.step.start() creates a starting formula with all selected Predictor Variables including factors and smoothed continuous variables.
}
\usage{
grasp.step.start(sX)
}
\arguments{
  \item{sX}{...}
}
\author{ Anthony.Lehmann@unige.ch }
\seealso{ grasp \code{\link{grasp}}, grasp.in \code{\link{grasp.in}}}
\keyword{models}
